requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-cosmos
azure-functions==1.12.0
azure-identity
azure-storage-blob==12.14.1
requests
bs4
xxhash
fastapi
